추옵 진행
추옵 진행하면서 스탯 퍼센트항목 추가

+++ Assets/3. Scripts/Charater/StatBar.cs
@@ -50,13 +50,13 @@ public class StatBar : MonoBehaviour
             StatBarImage.fillAmount = Mathf.Lerp(StatBarImage.fillAmount, currentFill, Time.deltaTime * lerpSpeed);
         }
     }
-    public void SetMax(float maxValue)
+    public void SetMax(int maxValue)
     {
         StatBarMaxValue = maxValue;
         StatBarText.text = currentValue + " / " + StatBarMaxValue;
         currentFill = currentValue / StatBarMaxValue;
     }
-    public void Initialize(float maxValue, float currentValue)
+    public void Initialize(int maxValue, int currentValue)
     {
         StatBarMaxValue = maxValue;
         StatBarCurrentValue = currentValue;
